- Added a casting from time_t to unsigned long in login_sql/login.c:1009. May not seem the most appropiate solution, but the date time "ban_until" is defined as an int in the tables anyway! - Added a check to prevent attempting to send data to disconnected chars (sd->fd == 0) in clif_send_sub. - Fixed players with no connection reporting themselves as "GM" on the console output.
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@3071 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/login_sql/login.c
@@ -1006,7 +1006,7 @@ int parse_fromchar(int fd){
 						charif_sendallwos(-1, buf, 11);
 					}
 					ShowNotice("Account: %d Banned until: %ld\n", acc, timestamp);
-					sprintf(tmpsql, "UPDATE `%s` SET `ban_until` = '%ld', `state`='7' WHERE `%s` = '%d'", login_db, timestamp, login_db_account_id, acc);
+					sprintf(tmpsql, "UPDATE `%s` SET `ban_until` = '%ld', `state`='7' WHERE `%s` = '%d'", login_db, (unsigned long)timestamp, login_db_account_id, acc);
 					// query
 					if (mysql_query(&mysql_handle, tmpsql)) {
 						ShowSQL("DB server Error - %s\n", mysql_error(&mysql_handle));
